Fix a bug in remove(), get one of the tests to run successfully
The target API supports a lot less than HashMap does, so much of the test needed to be removed (proxy classes, java.io.Serializable, and various other weird parts of Map aren't tested).
The bug in remove() had to do with me misunderstanding the backward-shift algorithm; it seems fine in all implementations now.