Brought in the post-PR changes, found and fixed a bug in containsValue()
This bug was caused by a post-PR change that didn't equivalently express the semantics of the for loop(s) in containsValue(). It also affected findKey().

The JSON stuff isn't back yet, and I'm not sure if it will return if I wind up using this repo to run tests on libGDX code.